Fix bug for rank about GridSearchCV #20852
To solve this bug I tried to understand the module.
Here the main goal is to get the best set of  parameter. In this for multi metric evaluation, stores the best_index, best_params_ and best_score_iff refit is the one of the scorer name.
For single metric evalution refit_metric is ‘score’.
Best score is defined on best index. After testing its better to go with a standard deviation rather than means to find best rank